+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/controller/dataaggregation/contributions/ContributionsCountController.java
@@ -1,9 +1,12 @@
 package org.heigit.ohsome.ohsomeapi.controller.dataaggregation.contributions;
 
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.heigit.ohsome.ohsomeapi.controller.ParameterDescriptions;
 import org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor;
 import org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse;
 import org.heigit.ohsome.ohsomeapi.output.Response;
@@ -17,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @Api(tags = "Contributions Count")
 @RestController
-@RequestMapping("/contributions/count")
+@RequestMapping("/contributions")
 public class ContributionsCountController {
 
   /**
@@ -27,18 +30,20 @@ public class ContributionsCountController {
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
    * @return {@link org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse
    *         DefaultAggregationResponse}
-   * @throws Exception thrown by
-   *         {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor#count(boolean) count}
+   * @throws Exception thrown by {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor
+   *         #count(boolean, boolean) count}
    */
   @ApiOperation(value = "Count of OSM contributions", nickname = "contributionsCount",
       response = DefaultAggregationResponse.class)
-  @RequestMapping(value = "", method = {RequestMethod.GET, RequestMethod.POST},
+  @RequestMapping(value = "/count", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "contributionType", value = ParameterDescriptions.CONTRIBUTION_TYPE,
+          defaultValue = "", paramType = "query", dataType = "string", required = false)})
   public Response contributionsCount(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    ContributionsExecutor executor =
-        new ContributionsExecutor(servletRequest, servletResponse, false);
-    return executor.count(false);
+    var executor = new ContributionsExecutor(servletRequest, servletResponse, false);
+    return executor.count(false, false);
   }
 
   /**
@@ -48,20 +53,69 @@ public class ContributionsCountController {
    * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
    * @return {@link org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse
    *         DefaultAggregationResponse}
-   * @throws Exception thrown by
-   *         {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor#count(boolean) count}
+   * @throws Exception thrown by {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor
+   *         #count(boolean, boolean) count}
    */
   @ApiOperation(
       value = "Density of OSM contributions (number of contributions divided by the "
           + "total area in square-kilometers)",
       nickname = "contributionsCountDensity", response = DefaultAggregationResponse.class)
-  @RequestMapping(value = "/density", method = {RequestMethod.GET, RequestMethod.POST},
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "contributionType", value = ParameterDescriptions.CONTRIBUTION_TYPE,
+          defaultValue = "", paramType = "query", dataType = "string", required = false)})
+  @RequestMapping(value = "/count/density", method = {RequestMethod.GET, RequestMethod.POST},
       produces = {"application/json", "text/csv"})
   public Response contributionsCountDensity(HttpServletRequest servletRequest,
       HttpServletResponse servletResponse) throws Exception {
-    ContributionsExecutor executor =
-        new ContributionsExecutor(servletRequest, servletResponse, true);
-    return executor.count(false);
+    var executor = new ContributionsExecutor(servletRequest, servletResponse, true);
+    return executor.count(false, false);
   }
 
+  /**
+   * Gives the density of latest OSM contributions.
+   *
+   * @param servletRequest <code>HttpServletRequest</code> of the incoming request
+   * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
+   * @return {@link org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse
+   *         DefaultAggregationResponse}
+   * @throws Exception thrown by {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor
+   *         #count(boolean, boolean) count}
+   */
+  @ApiOperation(
+      value = "Density of the latest OSM contributions (number of contributions divided by the "
+          + "total area in square-kilometers)",
+      nickname = "contributionsLatestCountDensity", response = DefaultAggregationResponse.class)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "contributionType", value = ParameterDescriptions.CONTRIBUTION_TYPE,
+          defaultValue = "", paramType = "query", dataType = "string", required = false)})
+  @RequestMapping(value = "/latest/count/density", method = {RequestMethod.GET, RequestMethod.POST},
+      produces = {"application/json", "text/csv"})
+  public Response contributionsLatestCountDensity(HttpServletRequest servletRequest,
+      HttpServletResponse servletResponse) throws Exception {
+    var executor = new ContributionsExecutor(servletRequest, servletResponse, true);
+    return executor.count(false, true);
+  }
+
+  /**
+   * Gives the count of latest OSM contributions.
+   *
+   * @param servletRequest <code>HttpServletRequest</code> of the incoming request
+   * @param servletResponse <code>HttpServletResponse</code> of the outgoing response
+   * @return {@link org.heigit.ohsome.ohsomeapi.output.DefaultAggregationResponse
+   *         DefaultAggregationResponse}
+   * @throws Exception thrown by {@link org.heigit.ohsome.ohsomeapi.executor.ContributionsExecutor
+   *         #count (boolean,boolean) count}
+   */
+  @ApiOperation(value = "Count of latest OSM contributions", nickname = "contributionsLatestCount",
+      response = DefaultAggregationResponse.class)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "contributionType", value = ParameterDescriptions.CONTRIBUTION_TYPE,
+          defaultValue = "", paramType = "query", dataType = "string", required = false)})
+  @RequestMapping(value = "/latest/count", method = {RequestMethod.GET, RequestMethod.POST},
+      produces = {"application/json", "text/csv"})
+  public Response contributionsLatestCount(HttpServletRequest servletRequest,
+      HttpServletResponse servletResponse) throws Exception {
+    var executor = new ContributionsExecutor(servletRequest, servletResponse, false);
+    return executor.count(false, true);
+  }
 }