+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/inputprocessing/ResourceParameters.java
@@ -44,8 +44,11 @@ public class ResourceParameters {
       possibleParams.add("properties");
       possibleParams.add("clipGeometry");
     }
-    // removing deprecated params from newer endpoint
     if (uri.contains("/contributions")) {
+      if (uri.contains("/count")) {
+        possibleParams.add("contributionType");
+      }
+      // removing deprecated params from newer endpoint
       possibleParams.remove("types");
       possibleParams.remove("keys");
       possibleParams.remove("values");